requirements.txt
requests
html-table-parser-python3
prettytable
flask
gunicorn
flask-talisman